clean up commented out plugins
Removed commented out plugins largely left over from openlaw-core repo.

ScalaStyle is also removed, we were not running it here, and in the
future we are more likely to use something like scalafmt.

sbt-scala-js-map was previously commented out when making this build,
as it causes issues in containers and after talking to David it is not
actually being used.

Added context to TODOs on plugins that are still open questions.