create docs folder and import wiki docs (#62)
* Relocate CoC and Contributing files to docs folder

GitHub should still autodetect them there

* migrate release process docs from private wiki

* introduce markdown style linting

(We should see this should fail in CI as files have not been updated to
match prettier style yet.)

* run style:fix, makes markdown files compliant

* restrict custom JS style overrides to our src

I believe some of these custom rules would create broken source if
applied to other JS files not being transpiled (e.g. config files in
the root) -- even though we are not running style linting on those now,
better to be explicit and make sure we dont cause problems for people
down the line if they add more files.

* move prettier config into package.json

this ensures it is always there in various CI environments (one less
file to forget to copy), and puts the configuration close to the run
scripts so the full picture can be seen at once.